Consolidate .Additional and .MoreCoverage test partials into base files
Merges every <test>.Additional.cs and <test>.MoreCoverage.cs partial back into
its base <test>.cs file across Numeric, Array, and IList test classes. The
fragmentation existed only because new tests were added in successive sessions;
collapsing them produces one test file per member, which matches the convention
of the rest of the suite.

For IListExtensionsTests_IndexOf / _LastIndexOf / _ReplaceAll the classes are
now declared without the `partial` keyword since all tests live in a single
file. NumericExtensionsTests and ArrayExtensionsTests remain `partial` because
their root files (NumericExtensionsTests.cs, ArrayExtensionsTests.cs) declare
them so and other member files extend the same class.

No test methods were added, removed, or renamed by this refactor.
