+++ Bodu.Core/test/Collections.Generic.Extensions/IListExtensions.LastIndexOf.cs
@@ -11,8 +11,19 @@ using System.Collections.ObjectModel;
 namespace Bodu.Collections.Generic.Extensions;
 
 [TestClass]
-public sealed partial class IListExtensionsTests_LastIndexOf
+public sealed class IListExtensionsTests_LastIndexOf
 {
+    /// <summary>
+    /// Provides multi-match scenarios for verifying that <c>LastIndexOf</c> always returns the last matching index.
+    /// </summary>
+    public static IEnumerable<object[]> LastMatchData =>
+    [
+        new object[] { new[] { 5, 1, 5, 5, 5 }, 4 },
+        new object[] { new[] { 5, 5, 5, 5, 1 }, 3 },
+        new object[] { new[] { 5, 1, 1, 1, 1 }, 0 },
+        new object[] { new[] { 5, 5 },          1 },
+    ];
+
     /// <summary>
     /// Verifies that <c>LastIndexOf</c> returns the index of the last matching element when one exists.
     /// </summary>
@@ -24,6 +35,20 @@ public sealed partial class IListExtensionsTests_LastIndexOf
         int index = list.LastIndexOf(x => x == 2);
 
         Assert.AreEqual(3, index);
+    }
+
+    /// <summary>
+    /// Verifies that <c>LastIndexOf</c> returns the last matching index when multiple elements satisfy the predicate.
+    /// </summary>
+    [TestMethod]
+    [DynamicData(nameof(LastMatchData))]
+    public void LastIndexOf_WhenMultipleMatches_ShouldReturnLastMatchingIndex(int[] data, int expected)
+    {
+        IList<int> list = new List<int>(data);
+
+        int index = list.LastIndexOf(x => x == 5);
+
+        Assert.AreEqual(expected, index);
     }
 
     /// <summary>
@@ -50,6 +75,29 @@ public sealed partial class IListExtensionsTests_LastIndexOf
         int index = list.LastIndexOf(_ => true);
 
         Assert.AreEqual(-1, index);
+    }
+
+    /// <summary>
+    /// Verifies that the no-arg <c>LastIndexOf</c> overload returns <c>-1</c> on a single-element list
+    /// when the only element does not match.
+    /// </summary>
+    [TestMethod]
+    public void LastIndexOf_WhenListHasSingleNonMatchingElement_ShouldReturnMinusOne()
+    {
+        IList<int> list = new List<int> { 7 };
+
+        Assert.AreEqual(-1, list.LastIndexOf(x => x == 5));
+    }
+
+    /// <summary>
+    /// Verifies that <c>LastIndexOf</c> with reference-type elements correctly locates the final occurrence.
+    /// </summary>
+    [TestMethod]
+    public void LastIndexOf_WithReferenceTypes_ShouldReturnLastMatchingIndex()
+    {
+        IList<string?> list = new List<string?> { null, "a", null, "b" };
+
+        Assert.AreEqual(2, list.LastIndexOf(x => x is null));
     }
 
     /// <summary>
@@ -90,6 +138,76 @@ public sealed partial class IListExtensionsTests_LastIndexOf
         int index = list.LastIndexOf(x => x == 1, 4, 5);
 
         Assert.AreEqual(0, index);
+    }
+
+    /// <summary>
+    /// Verifies that <c>LastIndexOf</c> with a ranged search returns the last matching index inside the window
+    /// when the suffix also contains matches.
+    /// </summary>
+    [TestMethod]
+    public void LastIndexOf_WithRange_WhenSuffixContainsMatches_ShouldReturnLastMatchInsideWindow()
+    {
+        IList<int> list = new List<int> { 5, 5, 5, 5, 5 };
+
+        int index = list.LastIndexOf(x => x == 5, 2, 2);
+
+        Assert.AreEqual(2, index);
+    }
+
+    /// <summary>
+    /// Verifies that <c>LastIndexOf</c> can locate a match at the exact first slot of the backward search range.
+    /// </summary>
+    [TestMethod]
+    public void LastIndexOf_WithRange_WhenMatchIsAtRangeStart_ShouldReturnIndex()
+    {
+        IList<int> list = new List<int> { 5, 0, 0, 0 };
+
+        int index = list.LastIndexOf(x => x == 5, 3, 4);
+
+        Assert.AreEqual(0, index);
+    }
+
+    /// <summary>
+    /// Verifies that <c>LastIndexOf</c> can locate a match at the exact startIndex (last slot of the
+    /// backward range).
+    /// </summary>
+    [TestMethod]
+    public void LastIndexOf_WithRange_WhenMatchIsAtStartIndex_ShouldReturnStartIndex()
+    {
+        IList<int> list = new List<int> { 0, 0, 0, 5 };
+
+        int index = list.LastIndexOf(x => x == 5, 3, 4);
+
+        Assert.AreEqual(3, index);
+    }
+
+    /// <summary>
+    /// Verifies that <c>LastIndexOf</c> with <paramref name="count"/> equal to <c>0</c> on a non-empty list
+    /// returns <c>-1</c> without invoking the predicate.
+    /// </summary>
+    [TestMethod]
+    public void LastIndexOf_WithRange_WhenCountIsZeroOnNonEmptyList_ShouldReturnMinusOne()
+    {
+        IList<int> list = new List<int> { 1, 2, 3 };
+        int predicateCalls = 0;
+
+        int index = list.LastIndexOf(_ => { predicateCalls++; return true; }, 2, 0);
+
+        Assert.AreEqual(-1, index);
+        Assert.AreEqual(0, predicateCalls);
+    }
+
+    /// <summary>
+    /// Verifies that <c>LastIndexOf</c> accepts <c>startIndex == -1</c> on an empty list and returns <c>-1</c>.
+    /// </summary>
+    [TestMethod]
+    public void LastIndexOf_WhenListIsEmptyAndStartIndexIsMinusOne_ShouldReturnMinusOne()
+    {
+        IList<int> list = new List<int>();
+
+        int index = list.LastIndexOf(x => x == 1, -1);
+
+        Assert.AreEqual(-1, index);
     }
 
     /// <summary>
@@ -149,19 +267,6 @@ public sealed partial class IListExtensionsTests_LastIndexOf
         {
             _ = list.LastIndexOf(x => x == 1, list.Count);
         });
-    }
-
-    /// <summary>
-    /// Verifies that <c>LastIndexOf</c> accepts <c>startIndex == -1</c> on an empty list and returns <c>-1</c>.
-    /// </summary>
-    [TestMethod]
-    public void LastIndexOf_WhenListIsEmptyAndStartIndexIsMinusOne_ShouldReturnMinusOne()
-    {
-        IList<int> list = new List<int>();
-
-        int index = list.LastIndexOf(x => x == 1, -1);
-
-        Assert.AreEqual(-1, index);
     }
 
     /// <summary>
