+++ Bodu.Core/test/Extensions/NumericExtensionsTests.RoundToSignificantDigits.cs
@@ -55,19 +55,53 @@ public partial class NumericExtensionsTests
     }
 
     /// <summary>
+    /// Verifies that <c>RoundToSignificantDigits</c> for <see cref="double"/> accepts the minimum valid digit
+    /// count of <c>1</c>.
+    /// </summary>
+    [TestMethod]
+    public void RoundToSignificantDigits_Double_WhenDigitsIsMinimumValid_ShouldRoundToOneDigit()
+    {
+        Assert.AreEqual(100.0, 123.45.RoundToSignificantDigits(1), 1e-9);
+        Assert.AreEqual(0.001, 0.00123.RoundToSignificantDigits(1), 1e-9);
+    }
+
+    /// <summary>
+    /// Verifies that <c>RoundToSignificantDigits</c> for <see cref="double"/> accepts the maximum valid digit
+    /// count of <c>15</c>.
+    /// </summary>
+    [TestMethod]
+    public void RoundToSignificantDigits_Double_WhenDigitsIsMaximumValid_ShouldRoundWithoutThrowing()
+    {
+        double value = 1.234567890123456;
+        double result = value.RoundToSignificantDigits(15);
+
+        Assert.AreEqual(value, result, 1e-14);
+    }
+
+    /// <summary>
     /// Verifies that <c>RoundToSignificantDigits</c> throws when the requested digit count is out of range.
     /// </summary>
     [DataTestMethod]
     [DataRow(0)]
     [DataRow(-1)]
     [DataRow(16)]
-    public void RoundToSignificantDigits_Double_WhenDigitsOutOfRange_ShouldThrowArgumentOutOfRangeException(int digits)
-    {
+    public void RoundToSignificantDigits_Double_WhenDigitsOutOfRange_ShouldThrowArgumentOutOfRangeException(int digits) =>
         Assert.ThrowsExactly<ArgumentOutOfRangeException>(() =>
         {
             _ = 123.45.RoundToSignificantDigits(digits);
         });
-    }
+
+    /// <summary>
+    /// Verifies that <c>RoundToSignificantDigits</c> for <see cref="double"/> rounds values midway between
+    /// representable significant points away from zero.
+    /// </summary>
+    [DataTestMethod]
+    [DataRow(1.5, 1, 2.0)]
+    [DataRow(2.5, 1, 3.0)]
+    [DataRow(-1.5, 1, -2.0)]
+    [DataRow(-2.5, 1, -3.0)]
+    public void RoundToSignificantDigits_Double_WhenAtMidpoint_ShouldRoundAwayFromZero(double value, int digits, double expected) =>
+        Assert.AreEqual(expected, value.RoundToSignificantDigits(digits), 1e-12);
 
     /// <summary>
     /// Verifies that <c>RoundToSignificantDigits</c> on <see cref="decimal"/> rounds the value as expected.
@@ -80,6 +114,16 @@ public partial class NumericExtensionsTests
     }
 
     /// <summary>
+    /// Verifies that <c>RoundToSignificantDigits</c> for <see cref="decimal"/> rounds negative values symmetrically.
+    /// </summary>
+    [TestMethod]
+    public void RoundToSignificantDigits_Decimal_WhenNegative_ShouldRoundSymmetrically()
+    {
+        decimal result = (-12345.6789m).RoundToSignificantDigits(3);
+        Assert.AreEqual(-12300m, result);
+    }
+
+    /// <summary>
     /// Verifies that the decimal overload returns zero unchanged.
     /// </summary>
     [TestMethod]
@@ -87,16 +131,33 @@ public partial class NumericExtensionsTests
         Assert.AreEqual(0m, 0m.RoundToSignificantDigits(5));
 
     /// <summary>
+    /// Verifies that <c>RoundToSignificantDigits</c> for <see cref="decimal"/> accepts the minimum valid digit count.
+    /// </summary>
+    [TestMethod]
+    public void RoundToSignificantDigits_Decimal_WhenDigitsIsMinimumValid_ShouldRoundToOneDigit() =>
+        Assert.AreEqual(100m, 123.45m.RoundToSignificantDigits(1));
+
+    /// <summary>
+    /// Verifies that <c>RoundToSignificantDigits</c> for <see cref="decimal"/> accepts the maximum valid digit count.
+    /// </summary>
+    [TestMethod]
+    public void RoundToSignificantDigits_Decimal_WhenDigitsIsMaximumValid_ShouldRoundWithoutThrowing()
+    {
+        decimal value = 1.5m;
+        decimal result = value.RoundToSignificantDigits(28);
+
+        Assert.AreEqual(value, result);
+    }
+
+    /// <summary>
     /// Verifies that the decimal overload throws when the requested digit count is out of range.
     /// </summary>
     [DataTestMethod]
     [DataRow(0)]
     [DataRow(29)]
-    public void RoundToSignificantDigits_Decimal_WhenDigitsOutOfRange_ShouldThrowArgumentOutOfRangeException(int digits)
-    {
+    public void RoundToSignificantDigits_Decimal_WhenDigitsOutOfRange_ShouldThrowArgumentOutOfRangeException(int digits) =>
         Assert.ThrowsExactly<ArgumentOutOfRangeException>(() =>
         {
             _ = 123.45m.RoundToSignificantDigits(digits);
         });
-    }
 }
